Divide workflow run logic into LocalExecutor
