+++ lib/tumugi/executor/local_executor.rb
@@ -1,0 +1,110 @@
+require 'much-timeout'
+require 'parallel'
+require 'retriable'
+
+require 'tumugi/error'
+
+module Tumugi
+  module Executor
+    class LocalExecutor
+      def initialize(dag, logger=nil, worker_num: 1)
+        @dag = dag
+        @logger = logger || Tumugi::Logger.instance
+        @options = { worker_num: worker_num }
+      end
+
+      def execute
+        setup_task_queue(@dag)
+        Parallel.each(dequeue_task, { in_threads: @options[:worker_num] }) do |t|
+          if t.requires_failed?
+            t.state = :requires_failed
+            @logger.info "#{t.state}: #{t.id} has failed requires task"
+            next
+          end
+
+          if !t.ready?
+            enqueu_task(t)
+            next
+          end
+
+          MuchTimeout.optional_timeout(task_timeout(t), Tumugi::TimeoutError) do
+            @logger.info "start: #{t.id}"
+            if t.completed?
+              t.state = :skipped
+              @logger.info "#{t.state}: #{t.id} is already completed"
+            else
+              @logger.info "run: #{t.id}"
+              t.state = :running
+
+              begin
+                Retriable.retriable retry_options do
+                  t.run
+                end
+                t.state = :completed
+                @logger.info "#{t.state}: #{t.id}"
+              rescue => e
+                t.state = :failed
+                @logger.info "#{t.state}: #{t.id}"
+                @logger.error "#{e.message}"
+                @logger.debug e.backtrace.join("\n")
+              end
+            end
+          end
+        end
+      end
+
+      private
+
+      def retry_options
+        {
+          tries: Tumugi.config.max_retry,
+          base_interval: Tumugi.config.retry_interval,
+          max_interval: Tumugi.config.retry_interval * Tumugi.config.max_retry,
+          max_elapsed_time: Tumugi.config.retry_interval * Tumugi.config.max_retry,
+          multiplier: 1.0,
+          rand_factor: 0.0,
+          on_retry: on_retry
+        }
+      end
+
+      def on_retry
+        Proc.new do |exception, try, elapsed_time, next_interval|
+          if next_interval
+            logger.error "#{exception.class}: '#{exception.message}' - #{try} tries in #{elapsed_time} seconds and #{next_interval} seconds until the next try."
+          else
+            logger.error "#{exception.class}: '#{exception.message}' - #{try} tries in #{elapsed_time} seconds. Task failed."
+          end
+        end
+      end
+
+      def task_timeout(task)
+        timeout = task.timeout || Tumugi.config.timeout
+        timeout = nil if !timeout.nil? && timeout == 0 # for backward compatibility
+        timeout
+      end
+
+      def setup_task_queue(dag)
+        @queue = Queue.new
+        dag.tsort.each do |t|
+          @queue << t
+        end
+        @queue
+      end
+
+      def dequeue_task
+        lambda {
+          begin
+            @queue.pop(true)
+          rescue ThreadError
+            Parallel::Stop
+          end
+        }
+      end
+
+      def enqueu_task(task)
+        sleep 1
+        @queue << task
+      end
+    end
+  end
+end
